Re-number the protocol versions to reflect the "Beta Stage" of the Mastercoin Specification / Implementations
It seems logical to inform new comers that the Mastercoin Protocol is currently in the Beta stage of its development. I propose we can easily accomplish this by simply re-numbering the versions of the protocol (both past and present) to version numbers below that of 1.0

For example version 0.1 version 0.19 and version 0.2 for past versions and 0.3 for the current version. This gives the protocol enough room for growth between now and when it does reach a 1.0 stage of development.