Class A Rule Simplification
Class A is only intended to provide backwards compatibility & is not
used in any of the respective wallet implementations or for any features
other than basic simple send.  Thus the effort being expended on Class A
does not correlate with its utility to the project and this pull is
intended to put Class A to bed so we can move onto other more useful
work.